Move data source to its own file
