+++ AuroraEditor/Base/NavigatorSidebar/Model/ProjectNavigator/ProjectNavigatorViewController.swift
@@ -7,8 +7,6 @@
 
 import SwiftUI
 
-private let dragType: NSPasteboard.PasteboardType = .fileURL
-
 /// A `NSViewController` that handles the **ProjectNavigator** in the **NavigatorSideabr**.
 ///
 /// Adds a ``outlineView`` inside a ``scrollView`` which shows the folder structure of the
@@ -16,6 +14,7 @@ private let dragType: NSPasteboard.PasteboardType = .fileURL
 final class ProjectNavigatorViewController: NSViewController {
 
     typealias Item = FileSystemClient.FileItem
+    let dragType: NSPasteboard.PasteboardType = .fileURL
 
     var scrollView: NSScrollView!
     var outlineView: NSOutlineView!
@@ -23,7 +22,7 @@ final class ProjectNavigatorViewController: NSViewController {
     /// Gets the folder structure
     ///
     /// Also creates a top level item "root" which represents the projects root directory and automatically expands it.
-    private var content: [Item] {
+    var content: [Item] {
         guard let folderURL = workspace?.fileSystemClient?.folderURL else { return [] }
         let children = workspace?.fileItems.sortItems(foldersOnTop: true)
         guard let root = try? workspace?.fileSystemClient?.getFileItem(folderURL.path) else { return [] }
@@ -247,87 +246,6 @@ final class ProjectNavigatorViewController: NSViewController {
             expandParent(item: parent)
         }
         outlineView.expandItem(item)
-    }
-}
-
-// MARK: - NSOutlineViewDataSource
-
-extension ProjectNavigatorViewController: NSOutlineViewDataSource {
-    func outlineView(_ outlineView: NSOutlineView, numberOfChildrenOfItem item: Any?) -> Int {
-        guard let workspace = self.workspace else { return 0 }
-        if let item = item as? Item {
-            return item.appearanceWithinChildrenOf(searchString: workspace.filter,
-                                                   ignoreDots: true,
-                                                   ignoreTilde: true)
-        }
-        return content.count
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, child index: Int, ofItem item: Any?) -> Any {
-        guard let workspace = self.workspace,
-              let item = item as? Item
-        else { return content[index] }
-
-        return item.childrenSatisfying(searchString: workspace.filter,
-                                       ignoreDots: true,
-                                       ignoreTilde: true)[index]
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, isItemExpandable item: Any) -> Bool {
-        if let item = item as? Item {
-            return item.children != nil
-        }
-        return false
-    }
-
-    // MARK: Drag and Drop
-    func outlineView(_ outlineView: NSOutlineView, pasteboardWriterForItem item: Any) -> NSPasteboardWriting? {
-        guard let fileItem = item as? FileItem else {
-            Log.error("Item \(item) is not file item")
-            return nil
-        }
-        let pboarditem = NSPasteboardItem()
-        pboarditem.setString(fileItem.url.path, forType: dragType)
-        return pboarditem
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, validateDrop info: NSDraggingInfo,
-                     proposedItem item: Any?, proposedChildIndex index: Int) -> NSDragOperation {
-        guard let targetFileItem = item as? FileItem,
-              let draggedString = info.draggingPasteboard.string(forType: dragType)
-        else { return NSDragOperation(arrayLiteral: [])}
-        let draggedURL = URL(fileURLWithPath: draggedString)
-
-        // if the item is being dragged onto an item that it contains, do not move it.
-
-        // the target must be a folder
-        // the target must not be within the dragged url
-        // the target's url and dragged url's parent must differ
-        guard targetFileItem.isFolder &&
-              !targetFileItem.url.path.hasPrefix(draggedString) &&
-              draggedURL.deletingLastPathComponent() != targetFileItem.url
-        else { return NSDragOperation(arrayLiteral: []) }
-
-        return NSDragOperation.move
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, acceptDrop info: NSDraggingInfo,
-                     item: Any?, childIndex index: Int) -> Bool {
-        // just reuse the validation function to check if the drop should happen
-        guard self.outlineView(outlineView, validateDrop: info,
-                               proposedItem: item, proposedChildIndex: index) == .move,
-              let targetFileItem = item as? FileItem,
-              let draggedString = info.draggingPasteboard.string(forType: dragType)
-        else { return false }
-        let draggedURL = URL(fileURLWithPath: draggedString)
-        do {
-            let fileName = draggedURL.lastPathComponent
-            try FileManager.default.moveItem(at: draggedURL, to: targetFileItem.url.appendingPathComponent(fileName))
-        } catch {
-            Log.error("Moving item \(draggedURL) to \(targetFileItem.url) failed: \(error)")
-            return false
-        }
-        return true
     }
 }
